add armor, fix itemNumChosen exception, and update InfoMenuPanel for not usable items

+++ src/main/kotlin/com/roguelike/InfoMenuPanel.kt
@@ -25,10 +25,13 @@ class InfoMenuPanel(private val gamePanel: GamePanel) : JPanel(), ActionListener
 
         bindKey(Keys.KEY_EXECUTE) {
             val itemNum = gamePanel.player.itemNumChosen
-            if (itemNum != null) {
+            val itemsCount = gamePanel.player.getCurrentItems().size
+            if (itemNum != null && itemNum < itemsCount) {
                 val itemToExecute = gamePanel.player.getCurrentItems()[itemNum]
                 itemToExecute.execute(gamePanel.player)
-                gamePanel.player.deleteItem(itemToExecute)
+                if (itemToExecute.isUsable) {
+                    gamePanel.player.deleteItem(itemToExecute)
+                }
             }
         }
 
@@ -81,15 +84,23 @@ class InfoMenuPanel(private val gamePanel: GamePanel) : JPanel(), ActionListener
         val items = gamePanel.player.getCurrentItems()
 
         var i = 0
+        val armor = gamePanel.player.armor
         for (itm in items) {
             if (i == gamePanel.player.itemNumChosen && gamePanel.player.itemNumChosen != null) {
                 val add = 6
-
-                g.color = Color(0, 0, 0, 210)
+                if (armor != null && armor == itm) {
+                    g.color = Color.green
+                } else {
+                    g.color = Color(0, 0, 0, 210)
+                }
                 g.fill3DRect((55 + i * 41) - add, 55 - add, 40 + add, 50 + add, true)
                 g.drawImage(itm.getPicture(), (55 + i * 41) - add, 55 - add, 40 + add, 40 + add, null)
             } else {
-                g.color = Color(222, 222, 222)
+                if (armor != null && armor == itm) {
+                    g.color = Color.green
+                } else {
+                    g.color = Color(222, 222, 222)
+                }
                 g.fill3DRect(55 + i * 41, 55, 40, 40, true)
                 g.drawImage(itm.getPicture(), 55 + i * 41, 55, 40, 40, null)
             }
